Added function data-retrieve and deployed it
