+++ src/data-collect/package.json
@@ -1,0 +1,17 @@
+{
+  "name": "data-collect",
+  "version": "1.0.0",
+  "description": "",
+  "main": "data-collect.js",
+  "scripts": {
+    "test": "echo \"Error: no test specified\" && exit 1"
+  },
+  "keywords": [],
+  "author": "",
+  "license": "ISC",
+  "dependencies": {
+    "@aws-sdk/client-s3": "^3.758.0",
+    "axios": "^1.8.3",
+    "cheerio": "^1.0.0"
+  }
+}
