+++ src/data-retrieve/package.json
@@ -1,0 +1,15 @@
+{
+  "name": "data-retrieve",
+  "version": "1.0.0",
+  "description": "",
+  "main": "data-retrieve.js",
+  "scripts": {
+    "test": "echo \"Error: no test specified\" && exit 1"
+  },
+  "keywords": [],
+  "author": "",
+  "license": "ISC",
+  "dependencies": {
+    "@aws-sdk/client-s3": "^3.758.0"
+  }
+}
