--- src/package.json
@@ -1,21 +1,0 @@
-{
-  "name": "seng3011_w18a_bravo",
-  "version": "1.0.0",
-  "description": "Bravo is a set of financial data analysis APIs created for the SENG3011 Event Intelligence Ecosystem project.",
-  "main": "index.js",
-  "scripts": {
-    "test": "echo \"Error: no test specified\" && exit 1"
-  },
-  "keywords": [],
-  "author": "",
-  "license": "ISC",
-  "dependencies": {
-    "@aws-sdk/client-s3": "^3.758.0",
-    "aws-sdk": "^2.1692.0",
-    "axios": "^1.8.3",
-    "cheerio": "^1.0.0",
-    "dotenv": "^16.4.7",
-    "express": "^4.21.2",
-    "jest": "^29.7.0"
-  }
-}
