+++ terraform/main.tf
@@ -73,19 +73,38 @@ resource "aws_iam_role_policy_attachment" "lambda_basic_execution" {
   policy_arn = "arn:aws:iam::aws:policy/service-role/AWSLambdaBasicExecutionRole"
 }
 
-# Lambda Function
+# Lambda Function for Data Collection
 resource "aws_lambda_function" "nba_scraper_lambda" {
   function_name = "nba-scraper-lambda"
-  handler       = "data-collect.handler"  # Update the handler to match your file and function
+  handler       = "data-collect.handler"  # Handler for data-collect.js
   runtime       = "nodejs18.x"
   role          = aws_iam_role.lambda_exec_role.arn
 
-  filename         = "lambda-deployment-package.zip"  # Path to the ZIP file in the root directory
-  source_code_hash = filebase64sha256("lambda-deployment-package.zip")
+  filename         = "lambda-deployment-package-collect.zip"  # Path to the ZIP file
+  source_code_hash = filebase64sha256("lambda-deployment-package-collect.zip")
 
   environment {
     variables = {
       S3_BUCKET_NAME = aws_s3_bucket.nba_prediction_bucket.bucket
+      S3_FILE_PREFIX = "nba-stats"  # Prefix for the S3 file path
+    }
+  }
+}
+
+# Lambda Function for Data Retrieval
+resource "aws_lambda_function" "nba_retriever_lambda" {
+  function_name = "nba-retriever-lambda"
+  handler       = "data-retrieve.handler"  # Handler for data-retrieve.js
+  runtime       = "nodejs18.x"
+  role          = aws_iam_role.lambda_exec_role.arn
+
+  filename         = "lambda-deployment-package-retrieve.zip"  # Path to the ZIP file
+  source_code_hash = filebase64sha256("lambda-deployment-package-retrieve.zip")
+
+  environment {
+    variables = {
+      S3_BUCKET_NAME = aws_s3_bucket.nba_prediction_bucket.bucket
+      S3_FILE_PREFIX = "nba-stats"  # Prefix for the S3 file path
     }
   }
 }
@@ -96,14 +115,14 @@ resource "aws_api_gateway_rest_api" "nba_prediction_api" {
   description = "API Gateway for NBA Prediction Lambda"
 }
 
-# API Gateway Resource
+# API Gateway Resource for Scrape
 resource "aws_api_gateway_resource" "scrape" {
   rest_api_id = aws_api_gateway_rest_api.nba_prediction_api.id
   parent_id   = aws_api_gateway_rest_api.nba_prediction_api.root_resource_id
   path_part   = "scrape"  # Resource path
 }
 
-# API Gateway Method
+# API Gateway Method for Scrape
 resource "aws_api_gateway_method" "scrape_method" {
   rest_api_id   = aws_api_gateway_rest_api.nba_prediction_api.id
   resource_id   = aws_api_gateway_resource.scrape.id
@@ -111,7 +130,7 @@ resource "aws_api_gateway_method" "scrape_method" {
   authorization = "NONE"  # Allow public access
 }
 
-# API Gateway Integration
+# API Gateway Integration for Scrape
 resource "aws_api_gateway_integration" "scrape_integration" {
   rest_api_id = aws_api_gateway_rest_api.nba_prediction_api.id
   resource_id = aws_api_gateway_method.scrape_method.resource_id
@@ -122,11 +141,47 @@ resource "aws_api_gateway_integration" "scrape_integration" {
   uri                     = aws_lambda_function.nba_scraper_lambda.invoke_arn
 }
 
-# Lambda Permission for API Gateway
-resource "aws_lambda_permission" "apigw_lambda" {
-  statement_id  = "AllowAPIGatewayInvoke"
+# API Gateway Resource for Retrieve
+resource "aws_api_gateway_resource" "retrieve" {
+  rest_api_id = aws_api_gateway_rest_api.nba_prediction_api.id
+  parent_id   = aws_api_gateway_rest_api.nba_prediction_api.root_resource_id
+  path_part   = "retrieve"  # Resource path
+}
+
+# API Gateway Method for Retrieve
+resource "aws_api_gateway_method" "retrieve_method" {
+  rest_api_id   = aws_api_gateway_rest_api.nba_prediction_api.id
+  resource_id   = aws_api_gateway_resource.retrieve.id
+  http_method   = "GET"  # HTTP method
+  authorization = "NONE"  # Allow public access
+}
+
+# API Gateway Integration for Retrieve
+resource "aws_api_gateway_integration" "retrieve_integration" {
+  rest_api_id = aws_api_gateway_rest_api.nba_prediction_api.id
+  resource_id = aws_api_gateway_method.retrieve_method.resource_id
+  http_method = aws_api_gateway_method.retrieve_method.http_method
+
+  integration_http_method = "POST"
+  type                    = "AWS_PROXY"
+  uri                     = aws_lambda_function.nba_retriever_lambda.invoke_arn
+}
+
+# Lambda Permission for API Gateway (Scrape)
+resource "aws_lambda_permission" "apigw_scrape_lambda" {
+  statement_id  = "AllowAPIGatewayInvokeScrape"
   action        = "lambda:InvokeFunction"
   function_name = aws_lambda_function.nba_scraper_lambda.function_name
+  principal     = "apigateway.amazonaws.com"
+
+  source_arn = "${aws_api_gateway_rest_api.nba_prediction_api.execution_arn}/*/*"
+}
+
+# Lambda Permission for API Gateway (Retrieve)
+resource "aws_lambda_permission" "apigw_retriever_lambda" {
+  statement_id  = "AllowAPIGatewayInvokeRetriever"
+  action        = "lambda:InvokeFunction"
+  function_name = aws_lambda_function.nba_retriever_lambda.function_name
   principal     = "apigateway.amazonaws.com"
 
   source_arn = "${aws_api_gateway_rest_api.nba_prediction_api.execution_arn}/*/*"
@@ -135,7 +190,10 @@ resource "aws_lambda_permission" "apigw_lambda" {
 # API Gateway Deployment
 resource "aws_api_gateway_deployment" "nba_prediction_deployment" {
   rest_api_id = aws_api_gateway_rest_api.nba_prediction_api.id
-  depends_on  = [aws_api_gateway_integration.scrape_integration]
+  depends_on  = [
+    aws_api_gateway_integration.scrape_integration,
+    aws_api_gateway_integration.retrieve_integration
+  ]
 }
 
 # API Gateway Stage
